move Pike to own dir
