+++ packages/quicktype-core/src/language/Pike/constants.ts
@@ -1,0 +1,54 @@
+export const keywords = [
+    "auto",
+    "nomask",
+    "final",
+    "static",
+    "extern",
+    "private",
+    "local",
+    "public",
+    "protected",
+    "inline",
+    "optional",
+    "variant",
+    "void",
+    "mixed",
+    "array",
+    "__attribute__",
+    "__deprecated__",
+    "mapping",
+    "multiset",
+    "object",
+    "function",
+    "__func__",
+    "program",
+    "string",
+    "float",
+    "int",
+    "enum",
+    "typedef",
+    "if",
+    "do",
+    "for",
+    "while",
+    "else",
+    "foreach",
+    "catch",
+    "gauge",
+    "class",
+    "break",
+    "case",
+    "const",
+    "constant",
+    "continue",
+    "default",
+    "import",
+    "inherit",
+    "lambda",
+    "predef",
+    "return",
+    "sscanf",
+    "switch",
+    "typeof",
+    "global"
+] as const;
