+++ packages/quicktype-core/src/language/Pike/index.ts
@@ -1,0 +1,2 @@
+export { PikeTargetLanguage, pikeOptions } from "./language";
+export { PikeRenderer } from "./PikeRenderer";
