+++ packages/quicktype-core/src/language/Pike/utils.ts
@@ -1,0 +1,7 @@
+import { funPrefixNamer } from "../../Naming";
+import { legalizeCharacters, isLetterOrUnderscoreOrDigit, makeNameStyle } from "../../support/Strings";
+
+const legalizeName = legalizeCharacters(isLetterOrUnderscoreOrDigit);
+export const enumNamingFunction = funPrefixNamer("enumNamer", makeNameStyle("upper-underscore", legalizeName));
+export const namingFunction = funPrefixNamer("genericNamer", makeNameStyle("underscore", legalizeName));
+export const namedTypeNamingFunction = funPrefixNamer("typeNamer", makeNameStyle("pascal", legalizeName));
